move properties to each connector
