+++ src/connector/src/kafka/source/reader.rs
@@ -22,12 +22,10 @@ use futures::StreamExt;
 use rdkafka::config::RDKafkaLogLevel;
 use rdkafka::consumer::{Consumer, DefaultConsumerContext, StreamConsumer};
 use rdkafka::{ClientConfig, Offset, TopicPartitionList};
-use risingwave_common::error::ErrorCode::InternalError;
-use risingwave_common::error::RwError;
 
 use crate::base::{SourceMessage, SplitReader};
 use crate::kafka::split::KafkaSplit;
-use crate::properties::KafkaProperties;
+use crate::kafka::KafkaProperties;
 use crate::{Column, ConnectorStateV2, SplitImpl};
 
 const KAFKA_MAX_FETCH_MESSAGES: usize = 1024;
@@ -75,7 +73,7 @@ impl SplitReader for KafkaSplitReader {
         let consumer: StreamConsumer = config
             .set_log_level(RDKafkaLogLevel::Info)
             .create_with_context(DefaultConsumerContext)
-            .map_err(|e| RwError::from(InternalError(format!("consumer creation failed {}", e))))?;
+            .map_err(|e| anyhow!("consumer creation failed {}", e))?;
 
         if let ConnectorStateV2::Splits(splits) = state {
             log::debug!("Splits for kafka found! {:?}", splits);
