+++ src/connector/src/kinesis/config.rs
@@ -14,7 +14,7 @@
 
 use std::collections::HashMap;
 
-use anyhow::Result;
+use anyhow::{anyhow, Result};
 use aws_config::default_provider::credentials::DefaultCredentialsChain;
 use aws_config::sts::AssumeRoleProvider;
 use aws_sdk_kinesis::Client;
@@ -22,11 +22,9 @@ use aws_types::credentials::SharedCredentialsProvider;
 use aws_types::region::Region;
 use http::Uri;
 use maplit::hashmap;
-use risingwave_common::error::ErrorCode::ProtocolError;
-use risingwave_common::error::RwError;
 use serde::{Deserialize, Serialize};
 
-use crate::KinesisProperties;
+use crate::kinesis::KinesisProperties;
 
 #[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
 pub struct AwsAssumeRole {
@@ -93,7 +91,7 @@ impl AwsConfigInfo {
         Ok(config_loader.load().await)
     }
 
-    pub fn build(properties: KinesisProperties) -> risingwave_common::error::Result<Self> {
+    pub fn build(properties: KinesisProperties) -> Result<Self> {
         let stream_name = properties.stream_name;
         let region = properties.stream_region;
 
@@ -106,7 +104,7 @@ impl AwsConfigInfo {
         );
         if access_key.is_some() ^ secret_key.is_some() {
             return Err(
-                RwError::from(ProtocolError("Both Kinesis credential access key and Kinesis secret key should be provided or not provided at the same time.".into()))
+                anyhow!("Both Kinesis credential access key and Kinesis secret key should be provided or not provided at the same time.")
             );
         } else if let (Some(access), Some(secret)) = (access_key, secret_key) {
             credentials = Some(AwsCredentials {
