+++ src/connector/src/kinesis/mod.rs
@@ -27,3 +27,27 @@ const KINESIS_CREDENTIALS_SECRET_ACCESS_KEY: &str = "kinesis.credentials.secret"
 const KINESIS_CREDENTIALS_SESSION_TOKEN: &str = "kinesis.credentials.session_token";
 const KINESIS_ASSUMEROLE_ARN: &str = "kinesis.assumerole.arn";
 const KINESIS_ASSUMEROLE_EXTERNAL_ID: &str = "kinesis.assumerole.external_id";
+
+use serde::Deserialize;
+
+pub const KINESIS_CONNECTOR: &str = "kinesis";
+
+#[derive(Clone, Debug, Deserialize)]
+pub struct KinesisProperties {
+    #[serde(rename = "kinesis.stream.name")]
+    pub stream_name: String,
+    #[serde(rename = "kinesis.stream.region")]
+    pub stream_region: String,
+    #[serde(rename = "kinesis.endpoint")]
+    pub endpoint: Option<String>,
+    #[serde(rename = "kinesis.credentials.access")]
+    pub credentials_access_key: Option<String>,
+    #[serde(rename = "kinesis.credentials.secret")]
+    pub credentials_secret_access_key: Option<String>,
+    #[serde(rename = "kinesis.credentials.session_token")]
+    pub session_token: Option<String>,
+    #[serde(rename = "kinesis.assumerole.arn")]
+    pub assume_role_arn: Option<String>,
+    #[serde(rename = "kinesis.assumerole.external_id")]
+    pub assume_role_externeal_id: Option<String>,
+}
